+++ classes/local/status_provider/mod_assign.php
@@ -18,7 +18,7 @@ namespace local_reminders\local\status_provider;
 
 use assign;
 use cm_info;
-use local_reminders\local\status_provider;
+use local_reminders\local\completion_status_provider;
 
 /**
  * Submission status provider for assignments.
@@ -28,7 +28,7 @@ use local_reminders\local\status_provider;
  * @copyright   2025 Catalyst IT Australia Pty Ltd
  * @license     https://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
-class mod_assign extends status_provider {
+class mod_assign extends completion_status_provider {
     /**
      * {@inheritDoc}
      *
@@ -54,13 +54,6 @@ class mod_assign extends status_provider {
             return $submission->status === ASSIGN_SUBMISSION_STATUS_SUBMITTED;
         }
 
-        switch (self::get_completion_state($userid, $cm)) {
-            case COMPLETION_COMPLETE:
-            case COMPLETION_COMPLETE_PASS:
-            case COMPLETION_COMPLETE_FAIL:
-                return true;
-            default:
-                return false;
-        }
+        return parent::is_completed($userid, $cm);
     }
 }
