+++ tests/local/status_provider/mod_assign_test.php
@@ -14,7 +14,8 @@
 // You should have received a copy of the GNU General Public License
 // along with Moodle.  If not, see <https://www.gnu.org/licenses/>.
 
-namespace local_reminders\local;
+namespace local_reminders\local\status_provider;
+
 
 defined('MOODLE_INTERNAL') || die();
 
@@ -22,6 +23,7 @@ global $CFG;
 require_once($CFG->dirroot . '/mod/assign/tests/generator.php');
 
 use advanced_testcase;
+use local_reminders\local\completion_status;
 use mod_assign_test_generator;
 
 /**
